requirements.txt
azure-functions
openai
azure-cosmos
azure-storage-blob
python-dotenv
